requirements.txt
alembic==1.16.2
av==13.1.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.3
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
glcontext==3.0.0
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
ManimPango==0.6.0
MarkupSafe==3.0.2
mdurl==0.1.2
moderngl==5.12.0
networkx==3.4.2
numpy==2.2.4
pillow==11.2.1
pycairo==1.28.0
pycparser==2.22
pydub==0.25.1
pyglet==2.1.3
pyglm==2.8.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.0
razorpay==1.4.2
requests==2.32.3
screeninfo==0.8.1
setuptools==78.1.0
skia-pathops==0.8.0.post2
soupsieve==2.6
SQLAlchemy==2.0.41
srt==3.5.3
stripe==12.2.0
svgelements==1.9.6
typing_extensions==4.13.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
